Cleaning up tenant deletion logic to enable tenant deletion using TenantMgtAdminService
Tenant deletion is by default switched off unless Tenant.TenantDelete property is setup in carbon.xml. This allows this logic not to impact existing behavior. Also, it is possible to enable/disable listener invocation on a tenant deletion. This is turned off by default as we have not tested the functionality. To turn listeners on, one must set Tenant.ListenerInvocationPolicy.InvokeOnDelete to true in carbon.xml. UMData cleanup is removed from tenant deletion.